requirements.txt
# ------------------------
# Core async / web
# ------------------------
anyio==4.11.0
h11==0.16.0
sniffio==1.3.1
httpcore==1.0.9
httpx==0.28.1
httptools==0.7.1
websockets==15.0.1
watchfiles==1.1.1
uvicorn==0.38.0

# ------------------------
# FastAPI stack (aligned)
# ------------------------
fastapi==0.122.0
starlette==0.50.0
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-multipart==0.0.20
email-validator==2.3.0

# ------------------------
# CLI / DX
# ------------------------
click==8.3.1
typer==0.20.0
rich==14.2.0
rich-toolkit==0.16.0
shellingham==1.5.4
colorama==0.4.6
rignore==0.7.6

# ------------------------
# HTTP / parsing
# ------------------------
requests==2.32.5
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11
certifi==2025.11.12

# ------------------------
# Google / GenAI
# ------------------------
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
httplib2==0.31.0
proto-plus==1.26.1

# protobuf pinned to SAFE version
protobuf==5.29.5

# ------------------------
# LLM / Transformers
# ------------------------
transformers==4.57.3
tokenizers==0.22.1
huggingface-hub==0.36.0
safetensors==0.7.0
sentencepiece==0.2.1
tiktoken==0.12.0

# ------------------------
# Local LLM (CPU/GPU safe)
# ------------------------
llama_cpp_python==0.3.16
gguf==0.17.1

# ------------------------
# ML / Numeric (ABI safe)
# ------------------------
numpy==2.3.5
scipy==1.16.3
scikit-learn==1.7.2
joblib==1.5.2
threadpoolctl==3.6.0
torch==2.9.1

# ------------------------
# RAG / Search / Web extraction
# ------------------------
sentence-transformers==5.1.2
tavily-python==0.7.13
trafilatura==2.0.0
jusText==3.0.2
courlan==1.3.2
htmldate==1.9.4
dateparser==1.2.2
tld==0.13.1

# ------------------------
# HTML / XML
# ------------------------
lxml==6.0.2
lxml_html_clean==0.4.3
babel==2.17.0

# ------------------------
# Utils / infra
# ------------------------
diskcache==5.6.3
filelock==3.20.0
fsspec==2025.10.0
packaging==25.0
tqdm==4.67.1
regex==2025.11.3
networkx==3.6
mpmath==1.3.0
sympy==1.14.0

# ------------------------
# Images / Docs
# ------------------------
pillow==12.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
MarkupSafe==3.0.3
Jinja2==3.1.6
Pygments==2.19.2

# ------------------------
# Config / time / env
# ------------------------
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
PyYAML==6.0.3

# ------------------------
# Monitoring / security
# ------------------------
sentry-sdk==2.46.0
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
dnspython==2.8.0
six==1.17.0
